+++ maxocontracts/blocks/__init__.py
@@ -1,14 +1,14 @@
 # Blocks subpackage for MaxoContracts
 from .condition import ConditionBlock
 from .action import ActionBlock
-from .gamma_protector import GammaProtectorBlock
+from .gamma_protector import WellnessProtectorBlock
 from .sdv_validator import SDVValidatorBlock
 from .reciprocity import ReciprocityBlock
 
 __all__ = [
     "ConditionBlock",
     "ActionBlock",
-    "GammaProtectorBlock",
+    "WellnessProtectorBlock",
     "SDVValidatorBlock",
     "ReciprocityBlock"
 ]